# openapi.yaml
openapi: 3.0.3
info:
  title: HIMATIKA Backend API
  description: Dokumentasi API backend HIMATIKA dalam format OpenAPI/Swagger.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /api/mathquiz/attempts/:
    get:
      tags:
        - Mathquiz
      summary: Riwayat attempt pengguna
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Riwayat attempt.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuizAttemptResponse'
    post:
      tags:
        - Mathquiz
      summary: Mulai attempt quiz
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MulaiAttemptRequest'
      responses:
        '201':
          description: Attempt berhasil dibuat.
  /api/mathquiz/{id}/submit/:
    patch:
      tags:
        - Mathquiz
      summary: Selesaikan attempt quiz
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitQuizRequest'
      responses:
        '200':
          description: Attempt selesai.
  /api/mathquiz/leaderboard/:
    get:
      tags:
        - Mathquiz
      summary: Leaderboard quiz
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Daftar leaderboard.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardResponse'
...
  /api/rama/form/:
    get:
      tags:
        - Rama
      summary: Ambil form RAMA
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Form RAMA.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RamaFormResponse'
  /api/rama/submit/:
    post:
      tags:
        - Rama
      summary: Submit RAMA
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RamaSubmitRequest'
      responses:
        '201':
          description: Submission berhasil dibuat.
  /api/rama/hasil/:
    get:
      tags:
        - Rama
      summary: Ambil hasil submission RAMA
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Hasil submission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RamaSubmissionResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    QuizAttemptResponse:
      type: object
      properties:
        id:
          type: integer
        quiz_judul:
          type: string
        skor:
          type: integer
        total_poin:
          type: integer
        status:
          type: string
        waktu_mulai:
          type: string
          format: date-time
        waktu_selesai:
          type: string
          format: date-time
    MulaiAttemptRequest:
      type: object
      required: [quiz_set]
      properties:
        quiz_set:
          type: integer
    SubmitQuizRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              question_id:
                type: integer
              jawaban:
                type: string
    LeaderboardResponse:
      type: object
      properties:
        username:
          type: string
        total_poin:
          type: integer
        total_quiz:
          type: integer
        total_benar:
          type: integer
...
    RamaFormResponse:
      type: object
      properties:
        sudah_submit:
          type: boolean
        semester:
          type: object
        kategori:
          type: array
          items:
            type: object
        submission:
          type: object
    RamaSubmitRequest:
      type: object
      required: [jawaban]
      properties:
        jawaban:
          type: array
          items:
            type: object
            properties:
              kategori:
                type: integer
              kepuasan:
                type: string
              isi_aspirasi:
                type: string
    RamaSubmissionResponse:
      type: object
      properties:
        semester:
          type: string
        tanggal_submit:
          type: string
          format: date-time
        status:
          type: string
        jawaban:
          type: array
          items:
            type: object